@@ -9,7 +9,7 @@ Gem::Specification.new do |s|
   s.summary     = "Full-stack web application framework."
   s.description = "Ruby on Rails is a full-stack web framework optimized for programmer happiness and sustainable productivity. It encourages beautiful code by favoring convention over configuration."
 
-  s.required_ruby_version     = ">= 2.3.0"
+  s.required_ruby_version     = ">= 2.4.1"
   s.required_rubygems_version = ">= 1.8.11"
 
   s.license = "MIT"
