@@ -144,15 +144,13 @@ class DateHelperTest < ActionView::TestCase
     rubinius_skip "Date is written in Ruby and relies on Fixnum#/"
     jruby_skip "Date is written in Ruby and relies on Fixnum#/"
 
-    klass = RUBY_VERSION > "2.4" ? Integer : Fixnum
-
     # Make sure that we avoid {Integer,Fixnum}#/ (redefined by mathn)
-    klass.send :private, :/
+    Integer.send :private, :/
 
     from = Time.utc(2004, 6, 6, 21, 45, 0)
     assert_distance_of_time_in_words(from)
   ensure
-    klass.send :public, :/
+    Integer.send :public, :/
   end
 
   def test_time_ago_in_words_passes_include_seconds
